+++ mdns_plugin/CHANGELOG.md
@@ -62,6 +62,15 @@ when TXT records are updated, the onServiceUpdated callback is not called
 on the Android platform, since the in-built libraries don't have this
 functionality available.
 
+## 1.0.1
+
+The plugin no longer resolves found services automatically but will only do so
+after "true" is returned from the delegate `onServiceFound`. In addition, the
+`startDiscovery` method now has an optional argument which is `enableUpdating`
+which when set to true will call the `onServiceUpdated` delegate method when
+TXT records are updated on the iOS platform. However, this doesn't yet work
+on Android.
+